Convert site server to typescript
