+++ site/.gitignore
@@ -51,3 +51,5 @@ public/sitemap.xml
 public/robots.txt
 
 tsconfig.tsbuildinfo
+
+/dist
