+++ site/package.json
@@ -2,9 +2,9 @@
     "name": "starter-site",
     "private": true,
     "scripts": {
-        "dev": "npm run intl:compile && run-s gql:types generate-block-types && NODE_OPTIONS='--inspect=localhost:9230' dotenv -e .env -e .env.local -e .env.secrets -e .env.site-configs -- node server.js",
-        "build": "npm run intl:compile && run-p gql:types generate-block-types && next build",
-        "serve": "NODE_ENV=production node server.js",
+        "dev": "npm run intl:compile && run-s gql:types generate-block-types && NODE_OPTIONS='--inspect=localhost:9230' dotenv -e .env -e .env.local -e .env.secrets -e .env.site-configs -- ts-node --project tsconfig.server.json server.ts",
+        "build": "npm run intl:compile && run-p gql:types generate-block-types && tsc --project tsconfig.server.json && next build",
+        "serve": "NODE_ENV=production node dist/server.js",
         "gql:types": "graphql-codegen",
         "gql:watch": "graphql-codegen --watch",
         "prelint": "npm run intl:compile && run-p gql:types generate-block-types",
