+++ site/tracing.ts
@@ -1,8 +1,8 @@
-const { diag, DiagConsoleLogger, DiagLogLevel } = require("@opentelemetry/api");
-const opentelemetry = require("@opentelemetry/sdk-node");
-const { getNodeAutoInstrumentations } = require("@opentelemetry/auto-instrumentations-node");
-const { OTLPTraceExporter } = require("@opentelemetry/exporter-trace-otlp-http");
-const { IncomingMessage } = require("http");
+import { diag, DiagConsoleLogger, DiagLogLevel } from "@opentelemetry/api";
+import { getNodeAutoInstrumentations } from "@opentelemetry/auto-instrumentations-node";
+import { OTLPTraceExporter } from "@opentelemetry/exporter-trace-otlp-http";
+import * as opentelemetry from "@opentelemetry/sdk-node";
+import { IncomingMessage } from "http";
 
 diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.INFO);
